筋トレの登録

+++ app/src/main/java/com/app/body_manage/common/Int.kt
@@ -1,7 +1,7 @@
 package com.app.body_manage.common
 
-fun Int?.toKcal(): String = if (this != null) "$this kcal" else "-"
+fun Long?.toKcal(): String = if (this != null) "$this kcal" else "-"
 fun Int?.toNumber(): String = "${this ?: 0} 個"
-fun Int?.toKg(): String = "$this kg"
-fun Int?.toCount(): String = "$this 回"
-fun Int?.toSet(): String = "$this セット"
+fun Long?.toKg(): String = "$this kg"
+fun Long?.toCount(): String = "$this 回"
+fun Long?.toSet(): String = "$this セット"
